tests: Allow more tests to be skipped.
Specifically tests which involve btrfs, because btrfs is
currently broken upstream (RHBZ#863978).

+++ tests/btrfs/test-btrfs-subvolume-default.pl
@@ -24,6 +24,9 @@ use warnings;
 use Sys::Guestfs;
 use Sys::Guestfs::Lib qw(feature_available);
 
+# Allow the test to be skipped since btrfs is often broken.
+exit 77 if $ENV{SKIP_TEST_BTRFS_SUBVOLUME_DEFAULT_PL};
+
 my $testimg = "test1.img";
 
 unlink $testimg;
